sempost and semwait index int -> long

+++ ImageStreamIO.h
@@ -28,10 +28,10 @@ int ImageStreamIO_createIm(IMAGE *image, const char *name, long naxis, uint32_t 
 long ImageStreamIO_read_sharedmem_image_toIMAGE(const char *name, IMAGE *image);
 
 
-int ImageStreamIO_semwait(IMAGE *image, int index);
+int ImageStreamIO_semwait(IMAGE *image, long index);
 
 
-int ImageStreamIO_sempost(IMAGE *image, int flag);
+int ImageStreamIO_sempost(IMAGE *image, long flag);
 
 
 #endif
